+++ CHANGELOG.md
@@ -7,6 +7,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- implement `CAA` record support [RFC 6844].
+
+[RFC 6844]: https://www.rfc-editor.org/rfc/rfc6844.html
+
 ## [0.23.0] - 2026-04-25
 
 ### Added